[add-react] add actions and reducers
