+++ app/javascript/actions/index.js
@@ -1,0 +1,1 @@
+export { getThings, getThingsSuccess } from './things';
