+++ app/javascript/configureStore.js
@@ -1,5 +1,6 @@
 import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
+import rootReducer from './reducers';
 
 const initialState = {
   things: [
@@ -9,15 +10,6 @@ const initialState = {
     },
   ],
 };
-
-function rootReducer(state, action) {
-  switch (action.type) {
-    case 'GET_THINGS_SUCCESS':
-      return { things: action.json.things };
-    default:
-      return state;
-  }
-}
 
 export default function configureStore() {
   const store = createStore(
