+++ app/javascript/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import ThingsReducer from './things';
+
+const rootReducer = combineReducers({
+  things: ThingsReducer,
+});
+
+export default rootReducer;
